Bug fixes, moved input checks out of T3

+++ Header Files/countdown.h
@@ -3,7 +3,7 @@
 
 #include <xc.h>
 
-extern volatile uint8_t countdown;
+extern volatile uint16_t countdown;
 
 
 void startTimer(uint8_t min, uint8_t sec);
